Modify signup form to user form to allow for reusablity
