+++ src/pages/ProfilePage.jsx
@@ -1,10 +1,11 @@
+import UserForm from "../components/UserForm";
 
 
 export default function ProfilePage(props){
 
     return(
-        <div className="text-white bg-grey-div mx-auto my-10 w-mobile-width h-mobile-height p-4 rounded-main-div shadow-mobile-shadow">
-            <h1>This is the profile page</h1>
+        <div className="text-white bg-grey-div mx-auto my-10 w-4/5 p-4 rounded-main-div shadow-mobile-shadow">
+            <UserForm />
         </div>
     )
 }
